requirements.txt
pytest
python-dotenv==0.19.0
Flask==2.2.0
better-profanity==0.7.0
flask-wtf==0.15.0
password-validator==1.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2